+++ editor/package.json
@@ -223,6 +223,7 @@
     "remix-react": "0.0.2",
     "reselect": "4.1.7",
     "resize-observer-polyfill": "1.5.0",
+    "sanitize-html": "2.11.0",
     "scheduler": "0.17.0",
     "semver": "7.3.2",
     "settle-promise": "1.0.0",
